refactored out base class
